move out SelectData component
